refactor op methods from routes->operation
